+++ NppCSharpPluginPack/Main.cs
@@ -462,6 +462,11 @@ The current scroll ratio is {Math.Round(scrollPercentage, 2)}%.
             foreach (string file in fileNames)
                 MessageBox.Show(file);
         }
+
+        /// <summary>
+        /// This method previously used the deprecated <see cref="NppMsg.NPPM_GETSESSIONFILES"/> message,
+        /// but it now uses the C# standard library to achieve the same effect.
+        /// </summary>
         static void GetSessionFileNamesDemo()
         {
             if (!Directory.Exists(PluginConfigDirectory) || !File.Exists(sessionFilePath))
@@ -471,22 +476,24 @@ The current scroll ratio is {Math.Round(scrollPercentage, 2)}%.
                     MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return;
             }
-            int nbFile = (int)Win32.SendMessage(PluginBase.nppData._nppHandle, (uint)NppMsg.NPPM_GETNBSESSIONFILES, 0, sessionFilePath);
 
-            if (nbFile < 1)
-            {
-                MessageBox.Show($"No valid session file at path \"{sessionFilePath}\" in order to point to a valid session file",
-                    "No valid session file",
-                    MessageBoxButtons.OK, MessageBoxIcon.Error);
-                return;
-            }
-            MessageBox.Show($"Number of session files: {nbFile}");
+            var sessionFileInfo = new FileInfo(sessionFilePath);
 
-            using (ClikeStringArray cStrArray = new ClikeStringArray(nbFile, Win32.MAX_PATH))
+            string sessionFileText;
+
+            using (StreamReader reader = sessionFileInfo.OpenText())
             {
-                if (Win32.SendMessage(PluginBase.nppData._nppHandle, (uint)NppMsg.NPPM_GETSESSIONFILES, cStrArray.NativePointer, sessionFilePath) != IntPtr.Zero)
-                    foreach (string file in cStrArray.ManagedStringsUnicode) MessageBox.Show(file);
+                sessionFileText = reader.ReadToEnd();
             }
+
+            var sessionFileNameRegex = new Regex("^[\\x20\\t]*<File.*filename=\"([^\"]+)", RegexOptions.Multiline);
+
+            var sessionFileNameMatches = sessionFileNameRegex.Matches(sessionFileText);
+
+            MessageBox.Show($"Number of session files: {sessionFileNameMatches.Count}");
+
+            foreach (Match mtch in sessionFileNameMatches)
+                MessageBox.Show(mtch.Groups[1].Value);
         }
         static void SaveCurrentSessionDemo()
         {