+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 2. Added the ability to [translate the plugin into other languages](/README.md#translating-your-plugin-to-another-language).
 3. Made it so all forms subclass a base class, making it easier to implement recommended methods.
 4. `Npp.TryGetLengthAsInt` and `Npp.TryGetText` methods, which gracefully handle files that are too large to put all their text in a string.
+5. `NotepadPPGateway.GetVisibleViews` method to quickly determine whether Notepad++ is in split-window mode and, if not, which view is currently visible.
 
 ### Fixed
 
@@ -39,7 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 5. Fix bug (related to fixed bug 4 above) where *all settings with non-integer floating-point values* would cause Notepad++ to crash on start-up if the localization used `,` as the decimal separator. 
 6. Fix bug with [makerelease.bat](/makerelease.bat) where it did not properly copy the dependency DLLs into the zip files.
 7. Fix bug when too-large integers are entered in the selection-remembering form.
-8. Fix potential crash when running `Get File Names Demo` plugin command, by [removing reference to dangerous `NPPM_GETOPENFILENAMES` command](https://community.notepad-plus-plus.org/post/98818).
+8. Fix potential crash when running `Get File Names Demo` and `Get Session File Names Demo` plugin commands, by [removing references to dangerous `NPPM_GETOPENFILENAMES` and `NPPM_GETSESSIONFILES` commands](https://community.notepad-plus-plus.org/post/98818).
 9. Fix bug where `NotepadPPGateway.GetOpenFileNames` would list a `new 1` file for an invisible view if only one view is open (i.e., if Notepad++ does not have a split window)
 
 ## [0.0.3] - 2024-02-26